requirements.txt
# Core dependencies
openpyxl>=3.1.0
pandas>=2.0.0
pyyaml>=6.0
click>=8.1.0
python-frontmatter>=1.0.0
markdown>=3.4.0
python-dateutil>=2.8.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
flake8>=6.0.0
